refactor(bin-designer): split CutoutsSection/geometry into focused modules
Decompose the 661-LOC geometry.ts into a slim facade plus three
focused sibling modules. Public API preserved via re-exports; no
behavior change (104 existing tests still pass).

- geometryResize.ts (201 LOC): MIN_CUTOUT_SIZE, StartRect interface,
  calculateCutoutResize, constrainGroupDrag, clampCornerRadius,
  getResizeCursor + RESIZE_CURSORS map.
- geometryAlignment.ts (153 LOC): AlignmentGuide / GUIDE_SNAP_THRESHOLD,
  findAlignmentGuides, distributeHorizontally, distributeVertically,
  centerInBin.
- geometryFlips.ts (161 LOC): flipCutoutHorizontal / flipCutoutVertical
  / flipSelectionHorizontal / flipSelectionVertical + path-point
  mirror helpers.
- geometry.ts (214 LOC): facade re-exports + rotation/bounds/snap
  primitives (rotatePoint, getRotatedBounds, clampRotationToBounds,
  computeBounds, getEffectiveBounds, clampPosition,
  getEffectiveWidth/Depth, snapToGrid, Bounds interface).
